access log and project issue fixing

+++ access_log.php
@@ -169,7 +169,7 @@ error_reporting(E_ALL ^ E_NOTICE);
                                                                 $hrs = $diff / 60;
                                                                 $mins = $hrs % 60;
                                                                 $hrs = $hrs / 60;
-                                                                $working_hours =  ((int)$hrs . "." . (int)$mins);
+                                                                $working_hours = (float) ((int)$hrs . "." . (int)$mins);
                                                             
                                                                 if(number_format($working_hours,2) >=7.40){
                                                                 $class = 'access_log_ok';
